Consolidated all actual tests into `tests.c`, with the individual files in `tests/modules`
