+++ tests/tests.c
@@ -2,88 +2,34 @@
 // Created by MacBook on 06.01.2025.
 //
 
-// tests
+#include "modules/test_arrayprintf.h"
+#include "modules/test_dict.h"
+#include "modules/test_dynstr.h"
+#include "modules/test_foreach.h"
+#include "modules/test_list.h"
+#include "modules/test_misc_c.h"
+#include "modules/test_queue.h"
+#include "modules/test_sets.h"
+#include "modules/test_string_utils.h"
+#include "modules/test_swex.h"
 
-#include "test_dict.h"
-#include "test_dynstr.h"
-#include "test_foreach.h"
-#include "test_list.h"
-#include "test_misc_c.h"
-#include "test_string_utils.h"
-#include "test_swex.h"
-
-
-void run_list( void )
-{
-    RUN_TEST( list_init );
-    RUN_TEST( list_basic );
-    RUN_TEST( list_advanced );
-}
-
-void run_dynstr( void )
-{
-    RUN_TEST( init_as );
-    RUN_TEST( slices );
-    RUN_TEST( append );
-    RUN_TEST( appendn );
-    RUN_TEST( appendf );
-}
-
-void run_foreach( void )
-{
-    RUN_TEST( foreach );
-    RUN_TEST( foreach_set );
-    RUN_TEST( foreach_queue );
-}
-
-void run_swex( void )
-{
-    RUN_TEST( swex );
-}
-
-void run_misc_c( void )
-{
-    RUN_TEST( misc_c );
-}
-
-Private TEST( array_sprintf )
-{
-    int array[ 10 ];
-    for ( size_t i = 0; i < countof( array ); ++i )
-        array[ i ] = ( int ) ( i * i ) * ( i % 2 == 0 ? -1 : 1 );
-
-    str_t str, snd;
-    array_sprintf_d( str, array, countof( array ), int, "%i", ", " );
-    array_sprintf( snd, array, countof( array ), int, "%i" );
-    PrintInColor( stdout, BACKGROUND_YELLOW, "%s\n", str );
-
-    UNIT_TEST( cmpeq( strcmp( str, "[ 0, 1, -4, 9, -16, 25, -36, 49, -64, 81 ]" ) ) );
-    UNIT_TEST( cmpeq( strcmp( str, snd ) ) );
-    free( str );
-    free( snd );
-
-    void *items_gen = array;
-    array_sprintf_d( str, items_gen, countof( array ), int, "%+02i", " | " );
-    PrintInColor( stdout, BACKGROUND_YELLOW, "%s\n", str );
-
-    UNIT_TEST( cmpeq( strcmp(
-            str, "[ +0 | +1 | -4 | +9 | -16 | +25 | -36 | +49 | -64 | +81 ]" ) ) );
-    free( str );
-}
-END_TEST
 
 int main( void )
 {
-    run_misc_c();
-    run_foreach();
-    run_swex();
-    run_string_utils();
-    run_dynstr();
-    run_list();
+    SET_UNIT_TEST_VERBOSITY( false );
+    RUNALL_STRING_UTILS();
 
-    RUN_TEST( dict_init );
+    RUNALL_DYNSTR();
+    RUNALL_LIST();
+    RUNALL_DICT();
+    RUNALL_SETS();
+    RUNALL_QUEUE();
 
-    RUN_TEST( array_sprintf );
+    RUNALL_MISC();
+
+    RUNALL_ARRAYPRINTF();
+    RUNALL_FOREACH();
+    RUNALL_SWEX();
 
     FINISH_TESTING();
 }
